feat: remove favs
